+++ CHANGELOG.md
@@ -1,8 +1,39 @@
+
 # 变更日志
+## [0.7.4] - wechatauto包
+### 变更
+- 修改crawl函数里面见名知意
+## [0.7.3] - wechatauto包
+### 修复
+- 修改,有的时候点击头像第一次个人资料弹不出来,重复点击2次
+### 变更
+- 添加了新功能,加入回复间隔时间，以及和当前时间间隔数，比如7天，可更改当前时间
+```python
+from bruce_li_tc.wechatauto.wechat_video_automator.bruce_uiauto.bruce_uiautomation import WeChatVideoCrawler,handle_single_comment
+# 1. 创建爬虫管理器实例
+crawler = WeChatVideoCrawler()
+# 2. 初始化（启动w、定位窗口等）
+crawler.initialize(back_click_model=True)
+# 3. 执行爬取！一句话搞定！
+crawler.set_comment_callback(handle_single_comment)
+video_list_data = crawler.crawl("关键字",
+                                comment_key=["评论中的关键字","评论中的关键字2","评论中的关键字3"],
+                                skip_comment=True,
+                                comment_list=["回复话1","回复话2","回复话3"],
+                                comment_day="7",
+                                comment_datetime="2025-9-26 15:05:00",
+                                Interval_count=5,
+                                Interval_seconds=1)
+print("video_list_data",video_list_data)
+```
+
+## [0.7.2] - wechatauto包
+### 修复
+- 加入回复条数触发等待时间
+
 ## [0.7.1] - wechatauto包
 ### 修复
 - 修复了导包的问题，0.7.0不可用
-
 
 ## [0.7.0] - wechatauto包
 ### 变更